requirements.txt
python-telegram-bot==20.6
flask==2.3.2
mysql-connector-python==8.3.0
bcrypt==4.1.2
python-dotenv==1.0.1
jdatetime==4.1.1
requests==2.32.3